icap-v7-master/classifier/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
psycopg2==2.9.3
gunicorn==20.1.0
pandas==1.4.0
requests==2.32.5
oracledb==1.2.2
XlsxWriter==3.0.6
openpyxl==3.0.10
django-import-export==2.8.0
black==23.7.0
isort==5.12.0
flake8==6.1.0
PyMuPDF==1.23.8
rapidfuzz==3.5.2
numpy==1.22.3
pillow==10.4.0
joblib==1.4.2
scikit-learn==1.0.2
spacy==3.7.5
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-core==1.0.1
langchain-groq==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
python-dotenv==1.1.1
pyyaml==6.0.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.25.0
opencv-python-headless==4.8.1.78 
onnxruntime==1.16.3 
onnxtr==0.8.0

# Message Queue
pika==1.3.2
# Redis for caching and pub/sub
redis[hiredis]==5.0.1